requirements.txt
asgiref==3.3.1
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
charset-normalizer==2.0.4
cryptography==3.4.6
dj-database-url==0.5.0
Django==3.1.7
django-environ==0.4.5
djangorestframework==3.12.2
google-api-core==1.31.1
google-auth==1.34.0
google-cloud-secret-manager==2.6.0
googleapis-common-protos==1.53.0
grpc-google-iam-v1==0.12.3
grpcio==1.39.0
idna==3.2
libcst==0.3.19
mypy-extensions==0.4.3
mysqlclient==2.0.3
packaging==21.0
proto-plus==1.19.0
protobuf==3.17.3
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyMySQL==1.0.2
pyparsing==2.4.7
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
rsa==4.7.2
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
typing-inspect==0.7.1
urllib3==1.26.6
uWSGI==2.0.18
whitenoise==5.3.0
